requirements.txt
pexpect
pyyaml
rich
dataclasses-json
